+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@ Release Notes
 =============
 
 # 0.4.0-SNAPSHOT (in progress)
+- fixed item sorting in case of integer/long or float/double clash
 - ...
 
 # 0.3.0